requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
streamlit==1.45.1
langchain==0.3.25
langchain-groq==0.3.2
langchain-experimental==0.3.4
pandas==2.2.3
matplotlib==3.10.3
seaborn==0.13.2
python-dotenv==1.1.0
requests==2.32.3
pydantic==2.11.3
tabulate==0.9.0
numpy==2.2.6
